imgui light control added + resize doesn't work :(
couldn't get resize to work rn but should not be much complicated since most of the hard setup is done. The textures recreate after nicely but the fence system might be broken.

also the compute shader doesn't display the galaxy shader now. I didn't mess up with it but it still is causing issues. Pretty sure the compute shader does run since when the OutputTexture is set to some color it does render it nicely, but the galaxy and stars output is not shown. Tried other shaders and same issue. Couldn't resolve it for now